# Conflicts:
#	appveyor.yml
#	assembly/pom.xml
#	common/kvstore/pom.xml
#	common/network-common/pom.xml
#	common/network-common/src/main/java/org/apache/spark/network/crypto/AuthEngine.java
#	common/network-common/src/main/java/org/apache/spark/network/crypto/TransportCipher.java
#	common/network-common/src/main/java/org/apache/spark/network/server/TransportRequestHandler.java
#	common/network-common/src/test/java/org/apache/spark/network/TransportClientFactorySuite.java
#	common/network-common/src/test/java/org/apache/spark/network/crypto/AuthEngineSuite.java
#	common/network-shuffle/pom.xml
#	common/network-shuffle/src/test/java/org/apache/spark/network/sasl/SaslIntegrationSuite.java
#	common/network-yarn/pom.xml
#	common/sketch/pom.xml
#	common/tags/pom.xml
#	common/unsafe/pom.xml
#	core/pom.xml
#	core/src/main/scala/org/apache/spark/HeartbeatReceiver.scala
#	core/src/main/scala/org/apache/spark/executor/Executor.scala
#	core/src/main/scala/org/apache/spark/internal/config/package.scala
#	core/src/main/scala/org/apache/spark/scheduler/EventLoggingListener.scala
#	core/src/main/scala/org/apache/spark/scheduler/TaskSchedulerImpl.scala
#	core/src/main/scala/org/apache/spark/scheduler/cluster/CoarseGrainedSchedulerBackend.scala
#	core/src/main/scala/org/apache/spark/status/AppStatusListener.scala
#	core/src/main/scala/org/apache/spark/status/AppStatusStore.scala
#	core/src/main/scala/org/apache/spark/storage/BlockManager.scala
#	core/src/main/scala/org/apache/spark/ui/JettyUtils.scala
#	core/src/main/scala/org/apache/spark/ui/WebUI.scala
#	core/src/main/scala/org/apache/spark/util/ClosureCleaner.scala
#	core/src/main/scala/org/apache/spark/util/Utils.scala
#	core/src/main/scala/org/apache/spark/util/io/ChunkedByteBuffer.scala
#	core/src/test/scala/org/apache/spark/ExecutorAllocationManagerSuite.scala
#	core/src/test/scala/org/apache/spark/SparkContextInfoSuite.scala
#	core/src/test/scala/org/apache/spark/deploy/SparkSubmitSuite.scala
#	core/src/test/scala/org/apache/spark/deploy/history/FsHistoryProviderSuite.scala
#	core/src/test/scala/org/apache/spark/security/CryptoStreamUtilsSuite.scala
#	core/src/test/scala/org/apache/spark/serializer/KryoSerializerSuite.scala
#	core/src/test/scala/org/apache/spark/status/AppStatusStoreSuite.scala
#	core/src/test/scala/org/apache/spark/storage/BlockManagerSuite.scala
#	core/src/test/scala/org/apache/spark/util/JsonProtocolSuite.scala
#	dev/create-release/release-build.sh
#	dev/create-release/spark-rm/Dockerfile
#	dev/deps/spark-deps-hadoop-2.6
#	dev/deps/spark-deps-hadoop-2.7
#	dev/deps/spark-deps-hadoop-3.1
#	dev/make-distribution.sh
#	docs/README.md
#	docs/configuration.md
#	docs/running-on-kubernetes.md
#	examples/pom.xml
#	external/avro/pom.xml
#	external/avro/src/test/scala/org/apache/spark/sql/avro/AvroSuite.scala
#	external/docker-integration-tests/pom.xml
#	external/flume-assembly/pom.xml
#	external/flume-sink/pom.xml
#	external/flume/pom.xml
#	external/kafka-0-10-assembly/pom.xml
#	external/kafka-0-10-sql/pom.xml
#	external/kafka-0-10-sql/src/main/scala/org/apache/spark/sql/kafka010/KafkaContinuousReader.scala
#	external/kafka-0-10/pom.xml
#	external/kafka-0-8-assembly/pom.xml
#	external/kafka-0-8/pom.xml
#	external/kinesis-asl-assembly/pom.xml
#	external/kinesis-asl/pom.xml
#	external/spark-ganglia-lgpl/pom.xml
#	graphx/pom.xml
#	hadoop-cloud/pom.xml
#	launcher/pom.xml
#	mllib-local/pom.xml
#	mllib/pom.xml
#	pom.xml
#	python/pyspark/context.py
#	repl/pom.xml
#	resource-managers/kubernetes/core/pom.xml
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/KubernetesUtils.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/features/BasicDriverFeatureStep.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/features/MountVolumesFeatureStep.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/submit/KubernetesClientApplication.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/deploy/k8s/submit/LoggingPodStatusWatcher.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/scheduler/cluster/k8s/KubernetesClusterSchedulerBackend.scala
#	resource-managers/kubernetes/core/src/main/scala/org/apache/spark/scheduler/cluster/k8s/KubernetesExecutorBuilder.scala
#	resource-managers/kubernetes/core/src/test/scala/org/apache/spark/deploy/k8s/features/BasicDriverFeatureStepSuite.scala
#	resource-managers/kubernetes/core/src/test/scala/org/apache/spark/scheduler/cluster/k8s/KubernetesExecutorBuilderSuite.scala
#	resource-managers/kubernetes/docker/src/main/dockerfiles/spark/Dockerfile
#	resource-managers/kubernetes/docker/src/main/dockerfiles/spark/entrypoint.sh
#	resource-managers/kubernetes/integration-tests/pom.xml
#	resource-managers/mesos/pom.xml
#	resource-managers/mesos/src/main/scala/org/apache/spark/scheduler/cluster/mesos/MesosCoarseGrainedSchedulerBackend.scala
#	resource-managers/yarn/pom.xml
#	resource-managers/yarn/src/main/scala/org/apache/spark/deploy/yarn/YarnAllocator.scala
#	resource-managers/yarn/src/main/scala/org/apache/spark/deploy/yarn/security/AMCredentialRenewer.scala
#	resource-managers/yarn/src/test/scala/org/apache/spark/deploy/yarn/ClientSuite.scala
#	resource-managers/yarn/src/test/scala/org/apache/spark/deploy/yarn/YarnAllocatorSuite.scala
#	sql/catalyst/pom.xml
#	sql/core/pom.xml
#	sql/core/src/main/java/org/apache/spark/sql/execution/datasources/parquet/VectorizedPlainValuesReader.java
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/csv/UnivocityParser.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcDeserializer.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcFileFormat.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcFilters.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcOutputWriter.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/datasources/orc/OrcUtils.scala
#	sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/continuous/ContinuousExecution.scala
#	sql/core/src/test/resources/sql-tests/results/describe-part-after-analyze.sql.out
#	sql/core/src/test/scala/org/apache/spark/sql/execution/WholeStageCodegenSuite.scala
#	sql/core/src/test/scala/org/apache/spark/sql/execution/columnar/InMemoryColumnarQuerySuite.scala
#	sql/core/src/test/scala/org/apache/spark/sql/execution/datasources/orc/OrcFilterSuite.scala
#	sql/hive-thriftserver/pom.xml
#	sql/hive/pom.xml
#	sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveShim.scala
#	sql/hive/src/main/scala/org/apache/spark/sql/hive/HiveStrategies.scala
#	sql/hive/src/main/scala/org/apache/spark/sql/hive/TableReader.scala
#	sql/hive/src/main/scala/org/apache/spark/sql/hive/client/HiveClientImpl.scala
#	sql/hive/src/main/scala/org/apache/spark/sql/hive/execution/CreateHiveTableAsSelectCommand.scala
#	sql/hive/src/test/scala/org/apache/spark/sql/hive/HiveExternalCatalogSuite.scala
#	streaming/pom.xml
#	tools/pom.xml

+++ core/src/main/java/org/apache/spark/unsafe/map/BytesToBytesMap.java
@@ -267,32 +267,45 @@ public final class BytesToBytesMap extends MemoryConsumer {
     }
 
     private void advanceToNextPage() {
-      synchronized (this) {
-        int nextIdx = dataPages.indexOf(currentPage) + 1;
-        if (destructive && currentPage != null) {
-          dataPages.remove(currentPage);
-          freePage(currentPage);
-          nextIdx --;
+      // SPARK-26265: We will first lock this `MapIterator` and then `TaskMemoryManager` when going
+      // to free a memory page by calling `freePage`. At the same time, it is possibly that another
+      // memory consumer first locks `TaskMemoryManager` and then this `MapIterator` when it
+      // acquires memory and causes spilling on this `MapIterator`. To avoid deadlock here, we keep
+      // reference to the page to free and free it after releasing the lock of `MapIterator`.
+      MemoryBlock pageToFree = null;
+
+      try {
+        synchronized (this) {
+          int nextIdx = dataPages.indexOf(currentPage) + 1;
+          if (destructive && currentPage != null) {
+            dataPages.remove(currentPage);
+            pageToFree = currentPage;
+            nextIdx--;
+          }
+          if (dataPages.size() > nextIdx) {
+            currentPage = dataPages.get(nextIdx);
+            pageBaseObject = currentPage.getBaseObject();
+            offsetInPage = currentPage.getBaseOffset();
+            recordsInPage = UnsafeAlignedOffset.getSize(pageBaseObject, offsetInPage);
+            offsetInPage += UnsafeAlignedOffset.getUaoSize();
+          } else {
+            currentPage = null;
+            if (reader != null) {
+              handleFailedDelete();
+            }
+            try {
+              Closeables.close(reader, /* swallowIOException = */ false);
+              reader = spillWriters.getFirst().getReader(serializerManager);
+              recordsInPage = -1;
+            } catch (IOException e) {
+              // Scala iterator does not handle exception
+              Platform.throwException(e);
+            }
+          }
         }
-        if (dataPages.size() > nextIdx) {
-          currentPage = dataPages.get(nextIdx);
-          pageBaseObject = currentPage.getBaseObject();
-          offsetInPage = currentPage.getBaseOffset();
-          recordsInPage = UnsafeAlignedOffset.getSize(pageBaseObject, offsetInPage);
-          offsetInPage += UnsafeAlignedOffset.getUaoSize();
-        } else {
-          currentPage = null;
-          if (reader != null) {
-            handleFailedDelete();
-          }
-          try {
-            Closeables.close(reader, /* swallowIOException = */ false);
-            reader = spillWriters.getFirst().getReader(serializerManager);
-            recordsInPage = -1;
-          } catch (IOException e) {
-            // Scala iterator does not handle exception
-            Platform.throwException(e);
-          }
+      } finally {
+        if (pageToFree != null) {
+          freePage(pageToFree);
         }
       }
     }
@@ -395,10 +408,12 @@ public final class BytesToBytesMap extends MemoryConsumer {
     }
 
     private void handleFailedDelete() {
-      // remove the spill file from disk
-      File file = spillWriters.removeFirst().getFile();
-      if (file != null && file.exists() && !file.delete()) {
-        logger.error("Was unable to delete spill file {}", file.getAbsolutePath());
+      if (spillWriters.size() > 0) {
+        // remove the spill file from disk
+        File file = spillWriters.removeFirst().getFile();
+        if (file != null && file.exists() && !file.delete()) {
+          logger.error("Was unable to delete spill file {}", file.getAbsolutePath());
+        }
       }
     }
   }
@@ -408,11 +423,11 @@ public final class BytesToBytesMap extends MemoryConsumer {
    *
    * For efficiency, all calls to `next()` will return the same {@link Location} object.
    *
-   * If any other lookups or operations are performed on this map while iterating over it, including
-   * `lookup()`, the behavior of the returned iterator is undefined.
+   * The returned iterator is thread-safe. However if the map is modified while iterating over it,
+   * the behavior of the returned iterator is undefined.
    */
   public MapIterator iterator() {
-    return new MapIterator(numValues, loc, false);
+    return new MapIterator(numValues, new Location(), false);
   }
 
   /**
@@ -422,19 +437,20 @@ public final class BytesToBytesMap extends MemoryConsumer {
    *
    * For efficiency, all calls to `next()` will return the same {@link Location} object.
    *
-   * If any other lookups or operations are performed on this map while iterating over it, including
-   * `lookup()`, the behavior of the returned iterator is undefined.
+   * The returned iterator is thread-safe. However if the map is modified while iterating over it,
+   * the behavior of the returned iterator is undefined.
    */
   public MapIterator destructiveIterator() {
     updatePeakMemoryUsed();
-    return new MapIterator(numValues, loc, true);
+    return new MapIterator(numValues, new Location(), true);
   }
 
   /**
    * Looks up a key, and return a {@link Location} handle that can be used to test existence
    * and read/write values.
    *
-   * This function always return the same {@link Location} instance to avoid object allocation.
+   * This function always returns the same {@link Location} instance to avoid object allocation.
+   * This function is not thread-safe.
    */
   public Location lookup(Object keyBase, long keyOffset, int keyLength) {
     safeLookup(keyBase, keyOffset, keyLength, loc,
@@ -446,7 +462,8 @@ public final class BytesToBytesMap extends MemoryConsumer {
    * Looks up a key, and return a {@link Location} handle that can be used to test existence
    * and read/write values.
    *
-   * This function always return the same {@link Location} instance to avoid object allocation.
+   * This function always returns the same {@link Location} instance to avoid object allocation.
+   * This function is not thread-safe.
    */
   public Location lookup(Object keyBase, long keyOffset, int keyLength, int hash) {
     safeLookup(keyBase, keyOffset, keyLength, loc, hash);
@@ -691,7 +708,10 @@ public final class BytesToBytesMap extends MemoryConsumer {
       assert (vlen % 8 == 0);
       assert (longArray != null);
 
-      if (numKeys == MAX_CAPACITY
+      // We should not increase number of keys to be MAX_CAPACITY. The usage pattern of this map is
+      // lookup + append. If we append key until the number of keys to be MAX_CAPACITY, next time
+      // the call of lookup will hang forever because it cannot find an empty slot.
+      if (numKeys == MAX_CAPACITY - 1
         // The map could be reused from last spill (because of no enough memory to grow),
         // then we don't try to grow again if hit the `growthThreshold`.
         || !canGrowArray && numKeys >= growthThreshold) {
@@ -738,10 +758,21 @@ public final class BytesToBytesMap extends MemoryConsumer {
         longArray.set(pos * 2 + 1, keyHashcode);
         isDefined = true;
 
-        if (numKeys >= growthThreshold && longArray.size() < MAX_CAPACITY) {
-          try {
-            growAndRehash();
-          } catch (OutOfMemoryError oom) {
+        // If the map has reached its growth threshold, try to grow it.
+        if (numKeys >= growthThreshold) {
+          // We use two array entries per key, so the array size is twice the capacity.
+          // We should compare the current capacity of the array, instead of its size.
+          if (longArray.size() / 2 < MAX_CAPACITY) {
+            try {
+              growAndRehash();
+            } catch (OutOfMemoryError oom) {
+              canGrowArray = false;
+            }
+          } else {
+            // The map is already at MAX_CAPACITY and cannot grow. Instead, we prevent it from
+            // accepting any more new elements to make sure we don't exceed the load factor. If we
+            // need to spill later, this allows UnsafeKVExternalSorter to reuse the array for
+            // sorting.
             canGrowArray = false;
           }
         }
@@ -886,6 +917,7 @@ public final class BytesToBytesMap extends MemoryConsumer {
     numKeys = 0;
     numValues = 0;
     freeArray(longArray);
+    longArray = null;
     while (dataPages.size() > 0) {
       MemoryBlock dataPage = dataPages.removeLast();
       freePage(dataPage);
